serving_container/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
xgboost==1.7.6
numpy==1.24.3
pydantic==2.5.0
google-cloud-storage==2.10.0
